[RPS-228] Keep maven dependencies up to date
Added a plugin that can be run to update all maven dependecy versions.
At the moment it is a manual command to be run but I plan to discuss
automating the running of this as part of the build.
I have run the tool and all dependencies are now up to date.
This involved removing the properties for versions from the root pom
as they will be explicitly updated by the tool now.

Also created a new property publication-system.basedir that will always
reference the root module base dir regardless for which module it is
evaluated in.