[RPS-266] Test for no PDF content in search
Added a test that searches for a term in a pdf, which I have added
to a pdf attachment to some documents in the test data, and checks
there are no results.

Also pushed some of the 'recent publications' into the future so
new test data does cause the publications overview test to fail.