[RPS-448] Remove MigratorImportHelper
Now that all the big migrations (Nestar, Goss, NIL) are over,
there is no need for migration-specific components to be included
in the codebase so they have been removed:
- MigrationImportHelper.java,
- MigratorImporterScript.groovy,
- MigratorImporterScript.yaml

Should they ever be needed, they can be recovered from Git history; the
last commit they are available in is bada740.

One of the code audits highlighted MigrationImportHelper not sanitising
paths in the ZIP files it handles and the decision was made that instead
of improving it, it's better to remove it now that it's no longer
needed. Since the aforementioned Groovy script relies on this class,
and is no longer needed either, it's been removed together with the
corresponding YAML config file.