[ADZ-1547] Custom, on-disk page caching
Implements on-disk caching for HTML rendered
for API Specifications.

Enables First Level Page Caching for the content
of /developer/api-catalogue folder and its sub-folders,
with the exclusion of API Catalogue page itself.

The underlying cache can be also used for other
kinds of 'heavy' content, i.e. one that takes long
to render and is infrequently accessed. The latter
use case is not supported by out of the box caching
solutions.