[DW-3690] Adds Accredited Stats Logo to Publications and Fixes Context (#2675)
This include the page hero and the results set on searches. They are
desplayed based on the parent collection. A tidy-up script was added
as some publications have leftover infomationTypes left over in the
database.

It also turned out that the getParentDocument method did not fetch
the correct parent series document when in the context of a Faceted
Results set. This commit includes a working new version and updates
the templates to use it. While doing this I noticed that
getParentDocument was being called many many times within a page load,
I have reduced this considerably.